Decoupling tl, gsi and rdb libraries from Qt bindings to enable lean strmrun application
